back_end/requirements.txt
Flask
pymongo
bcrypt
pyjwt
flask_cors
flask-talisman
gunicorn
pytest-flask
pytest
pydantic
python-dotenv